requirements.txt
streamlit==1.32.0
pandas==2.1.0
numpy==1.26.0
matplotlib==3.8.0
scikit-learn==1.3.0
pillow==10.0.0
mplsoccer==1.1.12
ipywidgets==8.1.1
